e2e rubrical CLI. /workspaces/rubrical/.venv/bin/activate
